requirements.txt
affine==2.4.0
altair==5.3.0
attrs==23.2.0
blinker==1.7.0
branca==0.7.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
contourpy==1.2.1
cycler==0.12.1
fiona==1.9.6
folium==0.16.0
fonttools==4.51.0
geographiclib==2.0
geonetworkx==0.5.3
geopandas==0.14.3
geopy==2.4.1
gitdb==4.0.11
GitPython==3.1.43
idna==3.6
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
networkx==3.2.1
nose==1.3.7
numpy==1.26.4
osmnx==1.9.1
packaging==24.0
pandas==2.2.1
pillow==10.3.0
protobuf==5.26.1
pyarrow==15.0.2
pydeck==0.8.1b0
Pygments==2.17.2
pyparsing==3.1.2
pyproj==3.6.1
python-dateutil==2.9.0.post0
pytz==2024.1
rasterio==1.3.9
referencing==0.34.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
scipy==1.13.0rc1
setuptools==69.2.0
shapely==2.0.3
shapelysmooth==0.1.1
six==1.16.0
smmap==6.0.0
snuggs==1.4.7
streamlit==1.33.0
streamlit-folium==0.18.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
xyzservices==2024.4.0
